+++ src/hooks/useDashboardStats.ts
@@ -7,6 +7,6 @@ export interface DashboardStats {
 }
 
 export const getDashboardStatus = async (): Promise<DashboardStats> => {
-    const { data } = await axios.get('https://api.example.com/dashboard');
+    const { data } = await axios.get('https://mocki.io/v1/71c8b321-2af4-48fe-82cb-2e45c0c7c976');
     return data;
 };
